requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-filebrowser==4.0.1
django-froala-editor==4.0.14
django-grappelli==3.0.3
django-tinymce==3.5.0
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.2
